requirements.txt
Click==7.0
Flask==1.0.2
httplib2==0.12.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
oauth2==1.9.0.post1
python-dotenv==0.10.1
splitwise==1.1.0
Werkzeug==0.14.1
